refactored folder structure
