+++ app.py
@@ -1,3 +1,3 @@
-from src.app import create_app
+from app import create_app
 
 app = create_app()
