+++ app/__init__.py
@@ -1,4 +1,4 @@
-from flask import Flask, url_for
+from flask import Flask
 from flask_cors import CORS
 
 def create_app():
