+++ app/blueprints/frontend/favicon/Favicon.py
@@ -1,5 +1,5 @@
 import os
-from flask import Blueprint, send_from_directory
+from flask import Blueprint, send_from_directory, current_app
 
 
 favicon_bp = Blueprint("favicon", __name__)
@@ -8,5 +8,4 @@ favicon_bp = Blueprint("favicon", __name__)
 @favicon_bp.route("/static/favico-16.ico", methods=["GET"])
 def favicon():
     """To resolve favicon 500 error and also to add image to the browser tab"""
-    from app import app
-    return send_from_directory(os.path.join(app.root_path, "static"), "favico-16.ico")
+    return send_from_directory(os.path.join(current_app.root_path, "static"), "favico-16.ico")
