--- deployment.yml
@@ -1,22 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: cloud-monitoring-app
-  labels:
-    app: cloud-monitoring-app
-spec:
-  replicas: 2
-  selector:
-    matchLabels:
-      app: cloud-monitoring-app
-  template:
-    metadata:
-      labels:
-        app: cloud-monitoring-app
-    spec:
-      containers:
-      - name: cloud-monitoring-app
-        image: 085606762228.dkr.ecr.us-west-2.amazonaws.com/cloud-monitoring-app:latest
-        imagePullPolicy: Always
-        ports:
-        - containerPort: 5000
