--- service.yml
@@ -1,12 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: cloud-monitoring-app-cluster
-spec:
-  selector:
-    app: cloud-monitoring-app
-  ports:
-    - protocol: TCP
-      port: 80
-      targetPort: 5000
-  type: LoadBalancer
