--- src/app/templates/index.html
@@ -1,120 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <!-- Required meta tags -->
-        <meta charset="utf-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1">
-        <!-- Bootstrap CSS -->
-        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
-        <!-- Custom CSS -->
-        <link rel="stylesheet" href="{{url_for('static', filename='css/index.css')}}">
-        <!-- Favicon 16*16 -->
-        <link type="image/png" sizes="16x16" rel="icon" href="{{ url_for('static', filename='favico-16.ico') }}">
-        <!-- SSE Data reception -->
-        <script type="text/javascript" src="{{ url_for('static', filename='js/index.js') }}"></script>
-        <title>Server Monitoring Metrics</title>
-    </head>
-    <body class="body">
-        <header>
-            <section class="curved">
-                <h1 class="header">Server Monitoring Metrics</h1>
-            </section>
-        </header>
-        <main>
-            <div class="row">
-                <div class="col">
-                    <div class="container-fluid p-2">
-                        <div class="card" style="width: 24rem;">
-                            <div class="card-body">
-                                <h5 class="card-title">CPU Usage</h5>
-                                <ul class="list-unstyled">
-                                    <ul>
-                                        <li>Percentage: <span id="cpu_usage_percentage"></span> %</li>
-                                        <li>Cores: <span id="cpu_count"></span></li>
-                                        <li>Frequency: <span id="cpu_frequency"></span> MHz</li>
-                                    </ul>
-                                </ul>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div class="col">
-                    <div class="container-fluid p-2">
-                        <div class="card">
-                            <div class="card-body">
-                                <h5 class="card-title">Memory Usage</h5>
-                                <ul class="list-unstyled">
-                                    <ul>
-                                        <li>Percentage: <span id="memory_usage_percentage"></span>%</li>
-                                        <li>Total: <span id="total_memory"></span> GB</li>
-                                        <li>Used: <span id="utilized_memory"></span> GB</li>
-                                        <li>Free: <span id="free_memory"></span> GB</li>
-                                        <li>Available: <span id="available_memory"></span> GB</li>
-                                    </ul>
-                                </ul>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div class="col">
-                    <div class="container-fluid p-2">
-                        <div class="card", style="width: 24rem;">
-                            <div class="card-body">
-                                <h5 class="card-title">Disk Usage</h5>
-                                <ul class="list-unstyled">
-                                    <ul>
-                                        <li>Total: <span id="total_disk_size"> GB</li>
-                                        <li>Usage: <span id="disk_utilized"></span> GB</li>
-                                        <li>Free: <span id="free_disk_space"></span> GB</li>
-                                        <li>Percentage: <span id="disk_usage_percentage"></span> %</li>
-                                    </ul>
-                                </ul>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div class="col">
-                    <div class="container-fluid p-2">
-                        <div class="card", style="width: 24rem;">
-                            <div class="card-body">
-                                <h5 class="card-title">Swap Memory</h5>
-                                <ul class="list-unstyled">
-                                    <ul>
-                                        <li>Total: <span id="total_swap_memory"></span> GB</li>
-                                        <li>Used: <span id="utilized_swap_memory"></span> GB</li>
-                                        <li>Free: <span id="free_swap_memory"></span> GB</li>
-                                        <li>Percentage: <span id="used_swap_memory_percentage"></span> %</li>
-                                    </ul>
-                                </ul>
-                            </div> 
-                        </div>
-                    </div>
-                </div>
-                <div class="col">
-                    <div class="container-fluid p-2">
-                        <div class="card", style="width: 24rem;">
-                            <div class="card-body">
-                                <h5 class="card-title">Network Usage</h5>
-                                <ul class="list-unstyled">
-                                    <ul>
-                                        <li>Bytes Sent: <span id="network_io_counters_bytes_sent"></span> Bytes</li>
-                                        <li>Bytes Received: <span id="network_io_counters_bytes_received"></span> Bytes</li>
-                                        <li>Packets Sent: <span id="network_io_counters_packets_sent"></span> Bytes</li>
-                                        <li>Packets Received: <span id="network_io_counters_packets_received"></span> Bytes</li>
-                                        <li>Errors In Receiving Packets: <span id="network_io_counters_errin"></span></li>
-                                        <li>Errors In Sending Packets: <span id="network_io_counters_errout"></span></li>
-                                        <li>Incoming Dropped Packets: <span id="network_io_counters_dropin"></span></li>
-                                        <li>Outgoing Dropped Packets: <span id="network_io_counters_dropout"></span></li>
-                                    </ul>
-                                </ul>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </main>
-        <!-- Optional JavaScript -->
-        
-        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ENjdO4Dr2bkBIFxQpeoTz1HIcje39Wm4jDKdf19U8gI4ddQ3GYNS7NTKfAdVQSZe" crossorigin="anonymous"></script>
-    </body>
-</html>
